Ev11model_Deployment/requirements.txt
streamlit==1.28.1
openai==1.3.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
sqlite3
bcrypt==4.0.1
plotly==5.17.0
altair==5.1.2
pillow==10.1.0
python-dotenv==1.0.0
google-generativeai==0.3.0
anthropic==0.7.10